fix: decompose AppDelegate and L10n to satisfy SwiftLint file/type length
Extract StatusBadge enum and drawBadge function into AppDelegate+StatusBadge.swift
to bring AppDelegate.swift from 523 lines to 471 (under the 500-line SwiftLint
--strict threshold). Also extracts inline support submenu construction into
buildSupportMenuItem() helper to keep configureStatusMenu() under 80 lines.
