+++ apps/macos-ui/Helm/AppDelegate.swift
@@ -354,29 +354,7 @@ private extension AppDelegate {
         settingsItem.submenu = settingsMenu
         menu.addItem(settingsItem)
 
-        let supportItem = NSMenuItem(
-            title: L10n.App.Settings.SupportFeedback.supportHelm.localized,
-            action: nil,
-            keyEquivalent: ""
-        )
-        let supportMenu = NSMenu()
-        let gitHubSponsorsItem = NSMenuItem(
-            title: L10n.App.Settings.SupportFeedback.gitHubSponsors.localized,
-            action: #selector(openGitHubSponsorsFromMenu),
-            keyEquivalent: ""
-        )
-        gitHubSponsorsItem.target = self
-        supportMenu.addItem(gitHubSponsorsItem)
-
-        let patreonItem = NSMenuItem(
-            title: L10n.App.Settings.SupportFeedback.patreon.localized,
-            action: #selector(openPatreonFromMenu),
-            keyEquivalent: ""
-        )
-        patreonItem.target = self
-        supportMenu.addItem(patreonItem)
-        supportItem.submenu = supportMenu
-        menu.addItem(supportItem)
+        menu.addItem(buildSupportMenuItem())
 
         menu.addItem(.separator())
 
@@ -403,6 +381,32 @@ private extension AppDelegate {
         statusItem?.menu = nil
         statusItem?.button?.menu = nil
         updateStatusMenuState()
+    }
+
+    private func buildSupportMenuItem() -> NSMenuItem {
+        let item = NSMenuItem(
+            title: L10n.App.Settings.SupportFeedback.supportHelm.localized,
+            action: nil,
+            keyEquivalent: ""
+        )
+        let submenu = NSMenu()
+        let gitHubSponsorsItem = NSMenuItem(
+            title: L10n.App.Settings.SupportFeedback.gitHubSponsors.localized,
+            action: #selector(openGitHubSponsorsFromMenu),
+            keyEquivalent: ""
+        )
+        gitHubSponsorsItem.target = self
+        submenu.addItem(gitHubSponsorsItem)
+
+        let patreonItem = NSMenuItem(
+            title: L10n.App.Settings.SupportFeedback.patreon.localized,
+            action: #selector(openPatreonFromMenu),
+            keyEquivalent: ""
+        )
+        patreonItem.target = self
+        submenu.addItem(patreonItem)
+        item.submenu = submenu
+        return item
     }
 
     func updateStatusMenuState() {
@@ -463,57 +467,5 @@ private extension AppDelegate {
 
     @objc func handleSystemAppearanceChanged() {
         updateStatusItemAppearance()
-    }
-}
-
-// MARK: - Status Badge
-
-private enum StatusBadge {
-    case count(Int, NSColor)
-    case symbol(String, NSColor)
-    case dot(NSColor)
-}
-
-private func drawBadge(_ badge: StatusBadge, in bounds: NSRect) {
-    switch badge {
-    case let .count(value, color):
-        let text = "\(value)"
-        let attributes: [NSAttributedString.Key: Any] = [
-            .font: NSFont.systemFont(ofSize: 7, weight: .bold),
-            .foregroundColor: NSColor.white
-        ]
-        let textSize = text.size(withAttributes: attributes)
-        let width = max(10, textSize.width + 5)
-        let badgeRect = NSRect(x: bounds.maxX - width - 1, y: bounds.maxY - 10, width: width, height: 10)
-        NSBezierPath(roundedRect: badgeRect, xRadius: 5, yRadius: 5).addClip()
-        color.setFill()
-        badgeRect.fill()
-        text.draw(
-            at: NSPoint(
-                x: badgeRect.minX + (badgeRect.width - textSize.width) / 2,
-                y: badgeRect.minY + (badgeRect.height - textSize.height) / 2 - 0.2
-            ),
-            withAttributes: attributes
-        )
-    case let .symbol(symbol, color):
-        let badgeRect = NSRect(x: bounds.maxX - 10, y: bounds.maxY - 10, width: 9, height: 9)
-        color.setFill()
-        NSBezierPath(ovalIn: badgeRect).fill()
-        let attributes: [NSAttributedString.Key: Any] = [
-            .font: NSFont.systemFont(ofSize: 7, weight: .bold),
-            .foregroundColor: NSColor.white
-        ]
-        let textSize = symbol.size(withAttributes: attributes)
-        symbol.draw(
-            at: NSPoint(
-                x: badgeRect.minX + (badgeRect.width - textSize.width) / 2,
-                y: badgeRect.minY + (badgeRect.height - textSize.height) / 2 - 0.4
-            ),
-            withAttributes: attributes
-        )
-    case let .dot(color):
-        let dotRect = NSRect(x: bounds.maxX - 7.5, y: bounds.maxY - 7.5, width: 5.5, height: 5.5)
-        color.setFill()
-        NSBezierPath(ovalIn: dotRect).fill()
     }
 }
